cambios sidebar

+++ database/migrations/2020_10_25_110337_create_table_gacondominio.php
@@ -41,8 +41,8 @@ class CreateTableGacondominio extends Migration
             $table->bigInteger('id_ctacontable');
             $table->string('nomctacontable', 100);
             $table->string('cod_ctacontable', 25);
-            $table->float('mo_ctacontable_bs',10,2); //10,2
-            $table->float('mo_ctacontable_ss',10,2);
+            $table->decimal('mo_ctacontable_bs',10,2); //10,2
+            $table->decimal('mo_ctacontable_ss',10,2);
             $table->timestamps();
         });
